backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
user-agents==2.2.0
qrcode==7.4.2
Pillow==10.0.0
requests==2.31.0
pydantic==2.5.0
sqlite3
python-multipart==0.0.6